feat: split editors on files AB#38298
